requirements.txt
# Core Dependencies
anthropic==0.40.0
pandas==2.2.0
numpy==1.26.4

# IBAN Validation
schwifty==2024.6.1

# Document Processing
PyPDF2==3.0.1
Pillow==10.4.0
pdf2image==1.17.0

# OCR (Optional - if using Tesseract)
pytesseract==0.3.13

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36

# Configuration Management
python-dotenv==1.0.1

# Security
cryptography==43.0.3

# HTTP Requests (for Azure integration)
requests==2.32.3
azure-storage-blob==12.23.1

# Utilities
python-dateutil==2.9.0
tqdm==4.66.5

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Logging
colorlog==6.9.0

# Email Notifications
secure-smtplib==0.1.1
